requirements.txt
openai
pandas
python-dotenv 
markitdown